move stateless to common pacakge
