--- dev/com.ibm.ws.microprofile.faulttolerance.cdi_fat/fat/src/com/ibm/websphere/microprofile/faulttolerance_fat/tests/stateless/circuitbreaker/TestException.java
@@ -1,16 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2025 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License 2.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *******************************************************************************/
-package com.ibm.websphere.microprofile.faulttolerance_fat.tests.stateless.circuitbreaker;
-
-public class TestException extends Exception {
-
-    private static final long serialVersionUID = 1L;
-
-}
